datapotal

+++ src/main/java/com/sens/pot/model/repository/PostsRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PostsRepository extends JpaRepository<Posts, Long>{
+public interface PostsRepository extends JpaRepository<Posts, Integer>{
     Optional<Posts> findById(Long id);
     List<Posts> findByCategory(Category category);
 }
